Add support for extension.json, refs 1367
