+++ extension.json
@@ -1,0 +1,28 @@
+{
+	"name": "SemanticMediaWiki",
+	"version": "3.0.0-alpha",
+	"author": [
+        "[http://korrekt.org Markus Krötzsch]",
+        "[https://www.mediawiki.org/wiki/User:Jeroen_De_Dauw Jeroen De Dauw]",
+        "James Hong Kong",
+        "[https://www.semantic-mediawiki.org/wiki/Contributors ...]"
+	],
+	"url": "https://www.semantic-mediawiki.org",
+	"descriptionmsg": "smw-desc",
+	"license-name": "GPL-2.0+",
+	"type": "semantic",
+	"requires": {
+		"MediaWiki": ">= 1.27"
+	},
+	"MessagesDirs": {
+		"SemanticMediaWiki": [
+			"i18n"
+		]
+	},
+	"callback": "SemanticMediaWiki::initExtension",
+	"ExtensionFunctions": [
+		"SemanticMediaWiki::onExtensionFunction"
+	],
+	"load_composer_autoloader":true,
+	"manifest_version": 1
+}
