+++ src/GlobalFunctions.php
@@ -216,6 +216,9 @@ function smwfGetLinker() {
 function enableSemantics( $namespace = null, $complete = false ) {
 	global $smwgNamespace;
 
+	// #1732 + #2813
+	wfLoadExtension( 'SemanticMediaWiki' );
+
 	// Apparently this is required (1.28+) as the earliest possible execution
 	// point in order for settings that refer to the SMW_NS_PROPERTY namespace
 	// to be available in LocalSettings
